Extracted the navbar avatar & name to fix the bug
