+++ client/src/components/navBar/navBar.css
@@ -1,6 +1,10 @@
 #navbar {
-  background-color: #EBA0AE;
-  background: linear-gradient(0deg, rgba(235,160,174,1) 0%, rgba(242,39,113,0.6) 100%);
+  background-color: #eba0ae;
+  background: linear-gradient(
+    0deg,
+    rgba(235, 160, 174, 1) 0%,
+    rgba(242, 39, 113, 0.6) 100%
+  );
   display: fixed;
   height: 55px;
   margin: -5px 0 0 0;
@@ -10,12 +14,12 @@
 }
 
 #logo {
- color: white;
- text-decoration: none;
- text-transform: uppercase;
- letter-spacing: 3px;
- margin-left: 10px;
- display: inline-block;
+  color: white;
+  text-decoration: none;
+  text-transform: uppercase;
+  letter-spacing: 3px;
+  margin-left: 10px;
+  display: inline-block;
 }
 
 a {
@@ -30,4 +34,17 @@ a {
 #nav-trips {
   color: inherit;
   text-decoration: none;
+}
+
+#nav-user-name {
+  float: left;
+  color: rgba(255, 255, 255, 0.6);
+  padding-top: 6px;
+  padding-right: 15px;
+  text-transform: capitalize;
+  font-size: 19px;
+}
+
+#nav-avatar {
+  margin-top: -4px;
 }
