+++ client/src/components/trips/trips.css
@@ -24,7 +24,6 @@
 
 .trip-body {
   justify-content: center;
-
 }
 
 .footer {
